Extract source-file scanning into a trait
